requirements_client.txt
# Client Dependencies
# Install with: pip install -r requirements_client.txt

# GUI and Display
tkinter>=3.11.0  # Usually comes with Python
numpy>=1.24.0

# Audio Capture
pyaudio>=0.2.13
pydub>=0.25.1

# Speech-to-Text (Local Whisper)
openai-whisper>=20240314

# Networking
requests>=2.31.0
aiohttp>=3.8.0

# Logging and Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: Cloud STT Providers (uncomment as needed)
# google-cloud-speech>=2.21.0  # Google Cloud Speech
# azure-cognitiveservices-speech>=1.31.0  # Azure Speech Services
